Fix warnings at compilation

+++ src/db/sqlite.cpp
@@ -514,7 +514,7 @@ std::string SqliteDbTable::getVar(const std::string& name)
     sqlite3_stmt *stmt;
     if (sqlite3_prepare(db, "SELECT value FROM vars WHERE name = ?", -1, &stmt, NULL) == SQLITE_OK)
     {
-        if (sqlite3_bind_text(stmt, 1, name.c_str(), name.length(), SQLITE_STATIC) == SQLITE_OK)
+        if (sqlite3_bind_text(stmt, 1, name.c_str(), static_cast<int>(name.length()), SQLITE_STATIC) == SQLITE_OK)
         {
             if((sqlite3_step(stmt) == SQLITE_ROW))
             {
@@ -546,9 +546,9 @@ bool SqliteDbTable::setVar(const std::string& name, const std::string& value)
 
     if (sqlite3_prepare(db, "INSERT OR REPLACE INTO vars (name, value) VALUES (?, ?)", -1, &stmt, NULL) == SQLITE_OK)
     {
-        if (sqlite3_bind_text(stmt, 1, name.c_str(), name.length(), SQLITE_STATIC) == SQLITE_OK)
+        if (sqlite3_bind_text(stmt, 1, name.c_str(), static_cast<int>(name.length()), SQLITE_STATIC) == SQLITE_OK)
         {
-            if (sqlite3_bind_blob(stmt, 2, value.data(), value.size(), SQLITE_STATIC) == SQLITE_OK)
+            if (sqlite3_bind_blob(stmt, 2, value.data(), static_cast<int>(value.size()), SQLITE_STATIC) == SQLITE_OK)
             {
                 if (sqlite3_step(stmt) == SQLITE_DONE)
                 {
